优化item组件, 删除todo前的状态确认

+++ src/components/Item/Item.jsx
@@ -21,7 +21,9 @@ export default class Item extends Component {
 
     headleDelete = id=>{
         // console.log(id)
-        this.props.deleteTodo(this.props.id);
+        if(window.confirm('确认删除吗?')){
+            this.props.deleteTodo(this.props.id);
+        }
     }
 
     render() {
